Refactor login and register component naming to promote consistency and clarity
